Change index buffer to uint32_t

+++ subprojects/render/include/render/extras.hpp
@@ -29,7 +29,7 @@ inline std::vector<Render::Vertex::Buffer> FillScreenVertices() {
     };
 }
 
-inline std::vector<uint16_t> FillScreenIndices() {
+inline std::vector<uint32_t> FillScreenIndices() {
     return {
         0, 1, 2,
         2, 3, 0,
